requirements.txt
# numpy==1.18.5
pyyaml==6.0
paho-mqtt==1.5.1
opencv-python-headless==4.5.2.52
webcolors==1.11.1